fix: Comment out unimplemented CQRS infrastructure and fix type issues
- Commented out references to non-existent CQRS infrastructure classes
- Commented out event and command handler registrations for now
- Fixed StablecoinAggregate::updatePosition to accept float parameters
- Updated CollateralPositionUpdated event to use float types
- Added TODO comments for future implementation

+++ app/Domain/Stablecoin/Aggregates/StablecoinAggregate.php
@@ -134,7 +134,7 @@ class StablecoinAggregate extends AggregateRoot
         return $this;
     }
 
-    public function updatePosition(int $collateral_amount, int $debt_amount, float $collateral_ratio): self
+    public function updatePosition(float $collateral_amount, float $debt_amount, float $collateral_ratio): self
     {
         $this->recordThat(
             new CollateralPositionUpdated(
